requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
atproto>=0.0.20
tweepy>=4.14.0